refact: rename the test files to Test1476
